widgets moved to seperate files and data moved to provider
