+++ patient/lib/presentation/assessments/models/assessment_card_model.dart
@@ -1,0 +1,13 @@
+class AssessmentCardModel {
+  final String id;
+  final String title;
+  final String description;
+  final String iconPath;
+
+  const AssessmentCardModel({
+    required this.id,
+    required this.title,
+    required this.description,
+    required this.iconPath,
+  });
+}
