+++ patient/lib/presentation/assessments/widgets/assessment_card.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+import 'package:patient/core/theme/theme.dart';
+import 'package:patient/presentation/assessments/models/assessment_card_model.dart';
+import 'package:patient/presentation/assessments/widgets/assessment_icon.dart';
+
+class AssessmentCard extends StatelessWidget {
+  final AssessmentCardModel assessment;
+  final VoidCallback onTap;
+
+  const AssessmentCard({
+    super.key,
+    required this.assessment,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onTap,
+      borderRadius: BorderRadius.circular(16),
+      child: Container(
+        decoration: BoxDecoration(
+          color: const Color(0xFFF9F9FC),
+          borderRadius: BorderRadius.circular(16),
+        ),
+        child: Padding(
+          padding: const EdgeInsets.only(top: 18.0, left: 18.0, right: 18.0),
+          child: Row(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      assessment.title,
+                      style: const TextStyle(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                        color: AppTheme.primaryColor,
+                      ),
+                    ),
+                    const SizedBox(height: 8),
+                    Text(
+                      assessment.description,
+                      style: const TextStyle(
+                        fontSize: 12,
+                        fontWeight: FontWeight.w500,
+                        color: Colors.black,
+                      ),
+                    ),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        AssessmentIcon(assessment: assessment),
+                      ],
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
