+++ patient/lib/provider/assessment_provider.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:patient/presentation/assessments/models/assessment_card_model.dart';
 import 'package:patient/repository/supabase_assessments_repository.dart';
 
 class AssessmentProvider with ChangeNotifier {
@@ -13,6 +14,28 @@ class AssessmentProvider with ChangeNotifier {
   Map<String, dynamic>? get assessment => _assessment;
   Map<int, String?> get selectedAnswers => _selectedAnswers;
 
+  final List<AssessmentCardModel> assessments = [
+    const AssessmentCardModel(
+      id: 'autism_quotient',
+      title: 'Autism Quotient(AQ)',
+      description: 'A quick referral guide to adults with suspected autism',
+      iconPath: 'assets/autism_icon.png',
+    ),
+    const AssessmentCardModel(
+      id: 'asrs_5',
+      title: 'ASRS-5',
+      description:
+          'The Adult ADHD Self-Report Scale for DSM-5 (ASRS-5) is a self-report screening scale for ADHD in adults',
+      iconPath: 'assets/adhd_icon.png',
+    ),
+    const AssessmentCardModel(
+      id: 'aq_10',
+      title: 'AQ-10',
+      description:
+          'The AQ-10 is a quick questionnaire that primary care practitioners can use to see if a person should be referred for an autism assessment.',
+      iconPath: 'assets/aq10_icon.png',
+    ),
+  ];
   Future<void> fetchAssessmentBySelectedId() async {
     try {
       //for testing purposes, the assessment id is hardcoded in the .env file and directly passing it if the assessment id is not set
